etl/requirements.txt
pandas==2.2.3
numpy==2.1.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
